Fix incorrect use of encryption keys and various size issues
Contrary to OpenSSL, the GP Internal API cannot encrypt with the
private key or decrypt with the public key of a RSA keypair [1].
Fix that.
Also fix several issues with passing buffer sizes. When
TEEC_MEMREF_WHOLE is used, the size received by secure world
is always the full size of the parent buffer. On return however,
the size may be updated with the value set by secure world [2].

[1] GP Internal Core API Specification v1.1.2 table 6-6
[2] GP Client API Specification v1.0 section 4.3.8
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>